refactor(knowledge): extract shared toRankableBill helper
CPD flagged the bill-row coercion as duplicated between
PersonalizedFeedService and PersonalizedRepActivityService (17 lines,
200 tokens — same drop-set logic for empty aiSummary, skip sentinel,
and missing topics/whoItAffects). Both services now import a single
toRankableBill function from personalized-feed/to-rankable-bill.ts.

No behavior change. Caught by the pre-push jscpd hook on the #769
branch.
